Set up db and basic API
Follows instructions here: https://howtojsonapi.com/rails.html to create
first iteration of the API. We now have a db with a table for words and
another for urls, and a one to many relationship between words and urls
so that we can have many urls for each word. The urls will be links to
videos of someone demonstrating the word in Makaton.