Create lookup table
Creates one table to act like a document store. The table is called
Lookups and it has two columns, word and video_url. Word is the primary
key. Word refers to the english word the user wants to find the Makaton
sign for, and video_url is an array of video links of people doing the
Makaton sign.

Command:
- `bin/rails g model lookups word:string:uniq:index video_url:text`

Then amended auto-generated migration file with:
- `id: false`
so this line became:
- `create_table :lookups, id: false do |t|`
This means rails won't set an automatic id and we will use the word
instead. I'm hoping the result will be that we can search a word
directly from the URL
Also amended file with:
- `, array: true, default: []`
so this line became:
- `t.text :video_url, array: true, default: []`
I'm hoping this will convert the video_url column to array type so that
we can store multiple urls here.

And finally amended model file with:
- `self.primary_key = 'word'`

Sources:
- https://stackoverflow.com/questions/10079195/how-to-set-primary-key-in-ruby-on-rails-migration
- https://howtojsonapi.com/rails.html