Make the main search page prettier
99% the work of chat GPT >>

Could you embed some CSS in this HTML file to make it look nicer and also work on mobile or web browser?

<!DOCTYPE html>
<html>
    <head>
    </head>
    <body>

        <input type="text" id="myText" value="Search">
        <button type="button" onclick="searchForWord()">Search</button>
        <h1 id="word"></h1>
        <iframe id="link" width="420" height="315"></iframe>

    </body>
    <script>
        function searchForWord() {
            let search_term = document.getElementById("myText").value;
            document.getElementById("word").innerHTML = search_term.toLowerCase();
            let link = document.getElementById("link")
            let word = document.getElementById("word")
            let request = new XMLHttpRequest();
            request.open("GET", `http://127.0.0.1:3000/search/${search_term}`);
            request.send();
            request.onload = () => {
                let response = JSON.parse(request.response)
                console.log(response)
                word.innerText = response.data.attributes["word"]
                link.src= response.data.attributes["video-url"][0];
            }
        }
    </script>
</html>
Sure! Here's an example of CSS that you can add to your HTML file to style the search form and the results:

css
Copy code
/* Set font family and size for the entire page */
body {
  font-family: Arial, sans-serif;
  font-size: 16px;
}

/* Style the search form */
form {
  display: flex;
  justify-content: center;
  align-items: center;
  margin: 50px 0;
}

input[type="text"] {
  font-size: 16px;
  padding: 10px;
  margin-right: 10px;
  width: 60%;
  border: 1px solid #ccc;
  border-radius: 4px;
}

button[type="submit"] {
  background-color: #4CAF50;
  color: white;
  font-size: 16px;
  padding: 10px 20px;
  border: none;
  border-radius: 4px;
  cursor: pointer;
}

button[type="submit"]:hover {
  background-color: #45a049;
}

/* Style the heading */
h1 {
  font-size: 32px;
  text-align: center;
  margin-top: 50px;
  margin-bottom: 30px;
}

/* Style the video iframe */
iframe {
  display: block;
  margin: 0 auto;
  width: 80%;
  height: 500px;
  border: none;
}
In the above example, the CSS code styles the search form by centering the input field and the button, adding some padding and border, and changing the background color and cursor on hover. The heading is also styled to be centered with a larger font size, and the video iframe is centered and given a fixed width and height.

To use this CSS in your HTML, you would need to add a <style> tag in the <head> section of your HTML and copy the CSS code into it. Here's an example:

html
Copy code
<!DOCTYPE html>
<html>
    <head>
        <title>Search for Makaton Signs</title>
        <style>
            /* Add the CSS code here */
        </style>
    </head>
    <body>
        <form>
            <input type="text" id="myText" value="Search">
            <button type="button" onclick="searchForWord()">Search</button>
        </form>
        <h1 id="word"></h1>
        <iframe id="link"></iframe>
        <script>
            function searchForWord() {
                let search_term = document.getElementById("myText").value;
                document.getElementById("word").innerHTML = search_term.toLowerCase();
                let link = document.getElementById("link")
                let word = document.getElementById("word")
                let request = new XMLHttpRequest();
                request.open("GET", `http://127.0.0.1:3000/search/${search_term}`);
                request.send();
                request.onload = () => {
                    let response = JSON.parse(request.response)
                    console.log(response)
                    word.innerText = response.data.attributes["word"]
                    link.src= response.data.attributes["video-url"][0];
                }
            }
        </script>
    </body>
</html>
With the updated HTML and CSS, the search form and the results will be styled to look nicer and work well on both mobile and web browsers.

Naomi Christie
The iframe is missing when I try this out.
I'm sorry to hear that. There could be a few reasons why the <iframe> is not displaying. Here are some things you can try to troubleshoot the issue:

Check the src attribute: Make sure that the src attribute of the <iframe> element is set to a valid URL. If the URL is incorrect or the video is not available, the iframe may not display.

Check the CSS: Make sure that the CSS you added to your HTML file is correctly targeting the <iframe> element. If the CSS is not applied properly, the iframe may not be visible.

Check the height and width: If the height or width attributes of the <iframe> element are set to 0 or a very small value, the iframe may not be visible. Make sure that the height and width are set to reasonable values to ensure that the iframe is visible.

Check the developer console: Use the developer console in your web browser to check for any errors or warnings that may be preventing the iframe from displaying. Look for any errors related to the <iframe> element or the video URL.

If none of these solutions work, please let me know and I'll try to help you further.

Naomi Christie
Access to XMLHttpRequest at 'http://127.0.0.1:3000/search/strawberry' from origin 'http://localhost:3000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
The error message you're seeing indicates that your web page is trying to make an HTTP request to a server running at http://127.0.0.1:3000 to fetch data, but the server is not set up to allow cross-origin requests from your web page running at http://localhost:3000.

Cross-origin requests are requests that are made from a different domain, protocol, or port than the one serving the page. Browsers enforce a security policy that blocks cross-origin requests by default to prevent malicious scripts from stealing user data.

To fix this error, you need to configure the server to allow cross-origin requests from your web page. You can do this by adding the rack-cors gem to your Rails application and configuring it to allow cross-origin requests.

Here's how you can add and configure the rack-cors gem:

Add the rack-cors gem to your Gemfile:
ruby
Copy code
gem 'rack-cors', require: 'rack/cors'
Run bundle install to install the gem.

Create a new file called cors.rb in your config/initializers directory with the following configuration:

ruby
Copy code
Rails.application.config.middleware.insert_before 0, Rack::Cors do
  allow do
    origins 'http://localhost:3000'
    resource '*', headers: :any, methods: [:get, :post, :options]
  end
end
In this configuration, we are allowing cross-origin requests from http://localhost:3000 to any resource (represented by the * wildcard), and we are allowing the GET, POST, and OPTIONS HTTP methods.

Restart your Rails server to pick up the changes.
With this configuration, your Rails server should now allow cross-origin requests from your web page, and the error should no longer occur.

Naomi Christie
Perfect.
Glad to hear that! If you have any further questions or issues, feel free to ask.