Use search box w/o needing to delete content first
In this updated code, we've added an event listener to the search input
that listens for the focus event, which is triggered when  the search
input is selected or clicked on, causing it to gain focus. When the
event is triggered, the event listener checks if the value of the input
is equal to the default value ("Search"), and if so, sets the value to
an empty string. This allows the user to start typing directly into the
search input without having to delete the default value.